back to not testing sdl

+++ sdl/render.c
@@ -37,16 +37,16 @@ void render(struct Ppu* ppu) {
 static void draw_ui_idle(struct AppState* state) {
   float scale = (float)win_width / (SCRN_WIDTH * 3);
   if (scale < 1.0f) scale = 1.0f;
-  int pw = (int)(300 * scale);
-  int row_h = (int)(30 * scale);
-  int label_h = (int)(16 * scale);
+  float pw = 300.f * scale;
+  float row_h = 30.f * scale;
+  float label_h = 16.f * scale;
   int recent = g_settings.recent_rom_count;
   int has_error = state->rom_error[0] != '\0';
-  int ph = row_h + 30 + (has_error ? label_h + 4 : 0) +
-           (recent > 0 ? label_h + 4 + recent * (row_h + 2) : 0);
-  int px = (win_width - pw) / 2;
-  int py = (win_height - ph) / 2;
-
+  float ph =
+      row_h + 30 + (has_error ? label_h + 4 : 0) +
+      ((float)recent > 0 ? label_h + 4 + (float)recent * (row_h + 2) : 0);
+  float px = ((float)win_width - pw) / 2.f;
+  float py = ((float)win_height - ph) / 2.f;
   if (nk_begin(ctx, "gbemu",
                nk_rect((float)px, (float)py, (float)pw, (float)ph),
                NK_WINDOW_TITLE | NK_WINDOW_BORDER | NK_WINDOW_NO_SCROLLBAR)) {
@@ -58,7 +58,6 @@ static void draw_ui_idle(struct AppState* state) {
       nk_label_colored(ctx, state->rom_error, NK_TEXT_LEFT,
                        nk_rgb(220, 80, 80));
     }
-
     if (recent > 0) {
       nk_layout_row_dynamic(ctx, label_h, 1);
       nk_label(ctx, "Recent:", NK_TEXT_LEFT);
